feat(PF-222): merchant merge (backend)

+++ src/main/java/com/mayureshpatel/pfdataservice/repository/merchant/query/MerchantQueries.java
@@ -88,5 +88,6 @@ public final class MerchantQueries {
     public static final String DELETE = """
             delete from merchants
             where id = :id
+              and user_id = :userId
             """;
 }
